improve requests logging
