+++ packages/bot-utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from './errors';
 export * from './flags';
 export * from './get';
 export * from './logger';
