+++ packages/bot-utils/src/utils.ts
@@ -1,5 +1,3 @@
-import { Response } from 'got';
-
 export const throwArgError = (invalid: string | undefined, type: string, expected: string) => {
 	throw new TypeError(`Invalid argument "${invalid}" for "${type}"!\nExpected "${expected}".`);
 };
@@ -7,52 +5,8 @@ export const throwArgError = (invalid: string | undefined, type: string, expecte
 export const throwOptError = (invalid: string | undefined, type: string, expected: string) => {
 	throw new TypeError(`Invalid option "${invalid}" for "${type}"!\nExpected "${expected}".`);
 };
-interface APIError {
-	error: {
-		message: string
-		type: 'validation_error' | 'no_probes_found' | 'api_error'
-		params?: {
-			[key: string]: string
-		}
-	}
-}
 
-export class PostError extends Error {
-	response: Response<unknown>;
 
-	location: string;
-
-	constructor(response: Response<unknown>, location: string, message = '') {
-		super(message);
-		this.message = message;
-		this.response = response;
-		this.location = location;
-	}
-}
-
-export const formatAPIError = (error: unknown): string => {
-	// @ts-ignore Discord error format
-	if (error.code === 50_001)
-		return 'Missing access! Please add the Globalping bot to this channel!';
-
-	if (error instanceof PostError) {
-		const { location, response } = error;
-		const { body } = response;
-		const errObj: APIError = JSON.parse(body as string) as APIError;
-		if (errObj.error.type === 'validation_error')
-			return `\`\`\`${errObj.error.message}\n${errObj.error.params ? Object.keys(errObj.error.params).map(key => `${errObj.error.params?.[key]}`).join('\n') : 'Unknown validation error.'}\`\`\`\nDocumentation and Support: https://github.com/jsdelivr/globalping`;
-
-		if (errObj.error.type === 'no_probes_found') {
-			return `\`\`\`${errObj.error.message} at location ${location}\`\`\`\nDocumentation and Support: https://github.com/jsdelivr/globalping`;
-		}
-		if (errObj.error.type === 'api_error') {
-			return `\`\`\`${errObj.error.message}\`\`\`\nDocumentation and Support: https://github.com/jsdelivr/globalping`;
-		}
-	} else if (error instanceof Error || error instanceof TypeError) {
-		return `\`\`\`${error.message}\`\`\`\nDocumentation and Support: https://github.com/jsdelivr/globalping`;
-	}
-	return `\`\`\`${error}\`\`\`\nDocumentation and Support: https://github.com/jsdelivr/globalping`;
-};
 
 interface Help {
 	[key: string]: {
